requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
mysql-connector-python==8.2.0

# Vector Database
chromadb==0.4.18

# ML Models
sentence-transformers==2.2.2
huggingface-hub==0.19.4
mlx-lm==0.8.0
torch>=2.0.0
numpy==1.24.3
scikit-learn==1.3.2

# Language Processing
langdetect==1.0.9
regex==2023.10.3

# Web Interface
jinja2==3.1.2
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0
pandas==2.1.4
#asyncio==3.4.3

# Development/Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0